Add validation of blocks
Added validation of blocks. First we validate the block itself,
comparing its own hash value with a new hash calculated in that moment.
Second, we check the current Block Previous Hash agains the Previous
Block Hash. If everything is OK, we return true.

Also added in the constructor this.hash = this.calculateHash(). No need
to create the Genesis Block with this line now, is automatically added
also to the Genesis Block